+++ fjs/cas/module.f.ts
@@ -9,7 +9,8 @@ import { join, normalize, parse } from '../path/module.f.mjs'
 import type { Vec } from '../types/bit_vec/types.ts'
 import { empty, length, maxLength, maxLengthBytes, msb, vec } from '../types/bit_vec/module.f.mjs'
 import { cBase32ToVec, vecToCBase32 } from '../basen/cbase32/module.f.mjs'
-import { foldStep, forEachStep, history, historyStep, okStep, pure, step, type Effect, type Operation } from '../effects/module.f.ts'
+import { foldStep, forEachStep, history, historyStep, okStep, pure, step } from '../effects/module.f.mjs'
+import type { Effect, Operation } from '../effects/types.ts'
 import { eff } from '../effects/eff/module.f.ts'
 import {
     access,