+++ fjs/mcp/cas/module.f.ts
@@ -103,7 +103,8 @@
  */
 import { string, option, or, boolean } from '../../types/rtti/module.f.mjs'
 import { stringify } from '../../media/json/module.f.ts'
-import { pure, step, type Effect } from '../../effects/module.f.ts'
+import { pure, step } from '../../effects/module.f.mjs'
+import type { Effect } from '../../effects/types.ts'
 import { type MemOp } from '../../effects/memory/module.f.ts'
 import { cBase32ToVec, vecToCBase32 } from '../../basen/cbase32/module.f.mjs'
 import { decode as base64Decode, encode as base64Encode } from '../../basen/base64/module.f.mjs'