+++ fjs/text/sgr/module.f.ts
@@ -9,7 +9,7 @@
 // https://en.wikipedia.org/wiki/ANSI_escape_code#C0_control_codes
 
 import { write, type Write, type WriteConsoles, type NodeProgramOptions } from '../../effects/node/module.f.ts'
-import { type Effect } from '../../effects/module.f.ts'
+import type { Effect } from '../../effects/types.ts'
 import { utf8 } from "../module.f.mjs"
 
 export const backspace: string = '\x08'